Dealing with Session Token
Based off the parse documentation, I created the class ParseErrorHandlingController to handle all parse errors from queries. Currently we're only dealing with the invalid session token query. I also added a call to get the current session in the background in the app delegate. If this fails with an invalid session token, the login screen is displayed.