Search is working - Project Organization
Search is now working. I found in the documentation that UISearchDisplayController was deprecated in iOS 8, so I migrated us to the UISearchController, which turned out to be overal fairly easy, except the scope segment control was covering the table view. So I modify the table view insets to compensate. It feels a bit like a hack, but I couldn't figure out a better way to change the table header frame on the fly.

I also did a bit of project cleanup. I hope I'm not stepping on your toes or making it more confusing. I just used a folder structure I've used before on past work with teams, and as projects get bigger with more files, it really helps a lot in finding what you are looking for.